create Bow Tie
As you can probably guess, Bow Ties is a lightweight bow tie management
application. You can switch between the different colors of bow tie you
own—the app assumes one of each—using the topmost segmented control—tap
“R” for red, “O” for orange and so on.
Tapping on a particular color pulls up an image of the tie and
populates several labels on the screen with specific information about
the tie. This includes:
• The name of the bow tie (so you can tell similarly-colored ones
apart); • The number of times you’ve worn the tie;
• The date you last wore the tie;
• Whether the tie is a favorite of yours.
The Wear button on the bottom-left increments the number of times
you’ve worn that particular tie and sets the “last worn” date to today.
Orange is not your color? Not to worry. The Rate button on the
bottom-right changes a bow tie’s rating. This particular rating system
uses a scale from 0 to 5, allowing for decimal values.